requirements.txt
aio-pika==5.6.0
attr==0.3.1
attrs==19.1.0
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
graphql-server-core==1.1.1
hypothesis==4.23.6
motor==2.0.0
marshmallow==2.19.2
marshmallow-annotations==2.4.0
pymongo==3.7.2
python-dotenv==0.10.3
sanic==19.3.1
Sanic-GraphQL==1.1.0
supervisor==4.0.3
